# app/static/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "description": "GT20 Team Final Project of Backend Track B-)",
        "version": "0.0.1",
        "title": "Startup Campus Backend Final Project",
        "contact": {
            "name": "Team GT20",
            "email": "gt20@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "url": "http://fashion-campus-gt20.my.id:5000"
        }
    ],
    "components": {
        "securitySchemes": {
            "Fashion-Campus-Auth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "in": "header",
                "name": "JWT-FC"
            }
        },
...
    },
    "paths": {
...
        "/categories": {
            "get": {
                "tags": [
                    "PRODUCT LIST"
                ],
                "summary": "Get Category for Searching Items",
                "responses": {
                    "200": {
                        "description": "Success Get Category",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "ADMIN PAGE"
                ],
                "summary": "Add category of product by Admin",
                "security": [
                    {
                        "Fashion-Campus-Auth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "string",
                                "properties": {
                                    "category_name": {
                                        "type": "string"
                                    }
                                }
                            },
                            "examples": {
                                "create_product": {
                                    "description": "Example of correct request body for creating product",
                                    "value": {
                                        "category_name": "Accessories"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Response for successfully create new product category",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "examples": {
                                    "create_product": {
                                        "description": "Response for successfully create new product category",
                                        "value": {
                                            "message": "Category added"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/UnauthorizedError"
                    }
                }
            }
        },
...
        "/cart": {
            "post": {
                "tags": [
                    "CART"
                ],
                "summary": "Add product to user's cart",
                "security": [
                    {
                        "Fashion-Campus-Auth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    },
                                    "quantity": {
                                        "type": "integer"
                                    },
                                    "size": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product added to cart",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/UnauthorizedError"
                    }
                }
            },
            "get": {
                "tags": [
                    "CART"
                ],
                "summary": "Get product's information in detail",
                "security": [
                    {
                        "Fashion-Campus-Auth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success get product's detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": {
                                                "type": "string"
                                            },
                                            "details": {
                                                "type": "object",
                                                "properties": {
                                                    "quantity": {
                                                        "type": "integer"
                                                    },
                                                    "size": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "price": {
                                                "type": "integer"
                                            },
                                            "image": {
                                                "type": "string"
                                            },
                                            "name": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/UnauthorizedError"
                    }
                }
            }
        },
...
        "/cart/{cart_id}": {
            "delete": {
                "tags": [
                    "CART"
                ],
                "summary": "Delete one product from user's cart",
                "security": [
                    {
                        "Fashion-Campus-Auth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "cart_id",
                        "description": "parameter where the id of product exist for deleting purpose",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success deleting product from cart",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "examples": {
                                    "success": {
                                        "description": "Cart deleted",
                                        "value": {
                                            "message": "Cart deleted"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/categories/{category_id}": {
            "put": {
                "tags": [
                    "ADMIN PAGE"
                ],
                "summary": "Updating existing certain category by Admin",
                "security": [
                    {
                        "Fashion-Campus-Auth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "category_id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "string",
                                "properties": {
                                    "category_name": {
                                        "type": "string"
                                    }
                                }
                            },
                            "examples": {
                                "updating category": {
                                    "description": "Example of correct request body for updating category",
                                    "value": {
                                        "category_name": "Hat"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Responses for successfully updating category",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "examples": {
                                    "updating category": {
                                        "description": "Example of successfully updating category by Admin",
                                        "value": {
                                            "message": "Category updated"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/UnauthorizedError"
                    }
                }
            },
            "delete": {
                "tags": [
                    "ADMIN PAGE"
                ],
                "summary": "Deleting certain category from Admin",
                "security": [
                    {
                        "Fashion-Campus-Auth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "category_id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Responses for successfully deleting category",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "examples": {
                                    "deleting category": {
                                        "description": "Example of successfully deleting category by Admin",
                                        "value": {
                                            "message": "Category deleted"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/UnauthorizedError"
                    }
                }
            }
        },
...
    }
}